requirements.txt
streamlit
langchain-chroma
langchain-huggingface
langchain-google-genai
langchain-core
python-dotenv
sentence-transformers
pymilvus>=2.3.0
numpy
datasets
pandas
scikit-learn
torch>=2.0.0
streamlit-audio-recorder
openai-whisper
esdk-obs-python>=3.22.0
requests>=2.31.0
flask>=2.3.0
sqlalchemy>=2.0.0
openpyxl>=3.1.0
psycopg2-binary>=2.9.0
pymysql>=1.1.0

